Added /token endpoint

See merge request rootteam/rootcause!153

+++ core/src/main/java/fr/uge/modules/linking/token/type/TypeDatetime.java
@@ -29,7 +29,7 @@ public class TypeDatetime implements TokenType{
         return TokenTypeId.ID_DATETIME.getId();
     }
 
-    public static Computation computeDateTimeProximity(Timestamp ldt1, Timestamp ldt2, float delta){
+    public static Computation computeDateTimeProximity(Timestamp ldt1, Timestamp ldt2, long delta){
         var time = (ldt1.getTime() - ldt2.getTime()) / 1000;
         return new Computation(new TypeDatetime(), ldt1.toString(), ldt2.toString(), fromTime(Math.abs(time), delta));
     }
